+++ src/model/UserInfo.ts
@@ -1,0 +1,11 @@
+interface UserInfo {
+  premium: boolean;
+  login: string;
+  email: string;
+  point: string;
+  premiumExpire: Date;
+  securityLock: boolean;
+  directDownload: boolean;
+  sslDownload: boolean;
+  token: string;
+}